% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwb.R
\docType{data}
\name{rwb}
\alias{rwb}
\title{Real Wage Bundle, USA 2014}
\format{
Consumption expenditure on the output of 53 industries.
}
\source{
\url{https://doi.org/10.34894/PJ2M1C}
}
\usage{
rwb
}
\description{
Personal Consumption Expenditure from the Input Output Table for the US economy in 2014.
}
\examples{
data(rwb)
}
\keyword{datasets}
