% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaiot.R
\docType{data}
\name{usaiot}
\alias{usaiot}
\title{USA IO Table}
\format{
Input Output table for 53 industries.
}
\source{
\url{https://doi.org/10.34894/PJ2M1C}
}
\usage{
usaiot
}
\description{
Input Output Table for the US economy from the World Input Output Database.
}
\examples{
data(usaiot)
}
\keyword{datasets}
