% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usarwb.R
\docType{data}
\name{usarwb}
\alias{usarwb}
\title{Real Wage Bundle, USA}
\format{
Consumption expenditure on the output of 53 industries.
}
\source{
\url{https://doi.org/10.34894/PJ2M1C}
}
\usage{
usarwb
}
\description{
Personal Consumption Expenditure from the Input Output Table for the US economy in 2014. This data is used to construct the real wage bundle for computing the price of production vector.
}
\examples{
data(usarwb)
}
\keyword{datasets}
